project_blackbird/requirements.txt
# Core (required to boot app)
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
pytest==8.3.2

# Optional (CV / reporting / AI / drone)
opencv-python==4.10.0.84
Pillow==10.4.0
reportlab==4.2.2
mavsdk==2.8.1
pymavlink==2.4.41
torch==2.4.0
torchvision==0.19.0
ultralytics==8.2.74
psycopg2-binary==2.9.9
